move open-file module into UI layer
